requirements.txt
Markdown==3.5.2
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
pyserial==3.5
qtwidgets==1.1
